+++ lib/data/app_providers/app_provider.dart
@@ -1,4 +1,4 @@
-import 'package:flutter_riverpod_template/data/remote/api/cient/api_config.dart';
+import 'package:flutter_riverpod_template/data/remote/api/client/api_config.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 final appProvider = Provider<ApiConfig>((ref) => ApiConfig('', apiKey: ''));
