+++ lib/data/remote/api/client/api_client.dart
@@ -1,17 +1,14 @@
+
 import 'package:dio/dio.dart';
-import 'package:flutter_riverpod_template/data/remote/api/cient/api_constants.dart';
+import 'package:flutter_riverpod_template/data/remote/api/client/api_constants.dart';
 import 'package:flutter_riverpod_template/feature/repository_list/models/repository_list_model.dart';
 import 'package:retrofit/retrofit.dart';
 
 part 'api_client.g.dart';
 
-
 @RestApi()
 abstract class ApiClient {
-  factory ApiClient(
-    Dio dio, {
-    String baseUrl,
-  }) = _ApiClient;
+  factory ApiClient(Dio dio, {String baseUrl}) = _ApiClient;
 
   @GET("${ApiPath.users}/{username}/repos")
   Future<List<RepositoryListModel>> getRepositories(
