+++ lib/data/remote/api/client/mock_api_client.dart
@@ -1,4 +1,4 @@
 import 'package:dio/dio.dart';
-import 'package:flutter_riverpod_template/data/remote/api/cient/api_client.dart';
+import 'package:flutter_riverpod_template/data/remote/api/client/api_client.dart';
 
 final mockApiClient = ApiClient(Dio(), baseUrl: "");
